fix : Scheduler 및 Transactional 충돌 오류 및 퀴리 오류 수정
